The greatest number (Puzzle by Magus)
Solution using class and object.
The puzzle is to print out the greatest number possible out of a input, which can include 0-9, '-' and '.'. Trailing/leading zeros must be removed, as well as dots ending a number.
The object newNumber() has all properties for evaluation of the input number, leading to the proper modifications.